# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Body of every non-2xx response (ADR 0006).",
        "properties": {
          "code": {
            "description": "Stable, machine-readable identifier in `snake_case`. Branch on this, never on `detail`.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "description": "Human-readable message. Safe to show; never contains internals.",
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present only for request-validation failures: one entry per offending field.",
            "title": "Errors"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FieldError": {
        "description": "One offending field of a request that failed validation.",
        "properties": {
          "loc": {
            "description": "Path to the field, starting with its location (`path`, `query`, `body`, ...).",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "message": {
            "description": "Human-readable reason the value was rejected.",
            "title": "Message",
            "type": "string"
          },
          "type": {
            "description": "Machine-readable identifier of the violated rule (e.g. `less_than_equal`).",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "message",
          "type"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "OperationKind": {
        "enum": [
          "tutor_activate",
          "student_activate",
          "student_stash",
          "student_pop",
          "student_deactivate",
          "tutor_deactivate"
        ],
        "title": "OperationKind",
        "type": "string"
      },
...
      "OperationStatus": {
        "enum": [
          "prepared",
          "committed",
          "expired",
          "failed",
          "cancelled"
        ],
        "title": "OperationStatus",
        "type": "string"
      },
...
      "TransitionCommitResponse": {
        "properties": {
          "committed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Committed At"
          },
          "kind": {
            "$ref": "#/components/schemas/OperationKind"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OperationStatus"
          }
        },
        "required": [
          "operation_id",
          "kind",
          "status",
          "committed_at"
        ],
        "title": "TransitionCommitResponse",
        "type": "object"
      },
      "TransitionPrepareResponse": {
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/OperationKind"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "plan": {
            "additionalProperties": true,
            "title": "Plan",
            "type": "object"
          }
        },
        "required": [
          "operation_id",
          "kind",
          "expires_at",
          "plan"
        ],
        "title": "TransitionPrepareResponse",
        "type": "object"
      },
      "TutorActivationCommitRequest": {
        "properties": {
          "category_channel_id": {
            "title": "Category Channel Id",
            "type": "integer"
          },
          "command_channel_id": {
            "title": "Command Channel Id",
            "type": "integer"
          }
        },
        "required": [
          "category_channel_id",
          "command_channel_id"
        ],
        "title": "TutorActivationCommitRequest",
        "type": "object"
      },
      "TutorActivationPrepareRequest": {
        "properties": {
          "guild_id": {
            "title": "Guild Id",
            "type": "integer"
          },
          "tutor_discord_id": {
            "title": "Tutor Discord Id",
            "type": "integer"
          }
        },
        "required": [
          "guild_id",
          "tutor_discord_id"
        ],
        "title": "TutorActivationPrepareRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2ClientCredentialsBearer": {
        "flows": {
          "clientCredentials": {
            "scopes": {
              "auth:clients:manage": "Manage application clients.",
              "bot:read": "Read bot API surface.",
              "bot:write": "Write bot API surface."
            },
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Backend of the skill-platform",
    "title": "skillforge",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/bot/tutors/activations/prepare": {
      "post": {
        "operationId": "bot_prepare_tutor_activation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorActivationPrepareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionPrepareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "invalid_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Invalid authentication credentials"
                    }
                  },
                  "missing_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Not authenticated"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "code": "forbidden",
                  "detail": "Not enough permissions"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing required scope: bot:write"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "transition_conflict": {
                    "value": {
                      "code": "transition_conflict",
                      "detail": "Transition conflict"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Transition conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "transition_validation": {
                    "value": {
                      "code": "transition_validation",
                      "detail": "Transition validation failed"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Transition validation failed"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentialsBearer": [
              "bot:write"
            ]
          }
        ],
        "summary": "Prepare Tutor Activation Endpoint",
        "tags": [
          "bot"
        ]
      }
    },
    "/api/v1/bot/tutors/activations/{operation_id}/commit": {
      "post": {
        "operationId": "bot_commit_tutor_activation",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorActivationCommitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionCommitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "invalid_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Invalid authentication credentials"
                    }
                  },
                  "missing_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Not authenticated"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "code": "forbidden",
                  "detail": "Not enough permissions"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing required scope: bot:write"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "operation_not_found": {
                    "value": {
                      "code": "operation_not_found",
                      "detail": "Operation not found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operation not found"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "operation_not_pending": {
                    "value": {
                      "code": "operation_not_pending",
                      "detail": "Operation is not in a prepared state (already committed, failed, or expired)"
                    }
                  },
                  "transition_conflict": {
                    "value": {
                      "code": "transition_conflict",
                      "detail": "Transition conflict"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operation is not in a prepared state (already committed, failed, or expired) / Transition conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "code": "validation_error",
                  "detail": "Request validation failed",
                  "errors": [
                    {
                      "loc": [
                        "query",
                        "limit"
                      ],
                      "message": "Input should be less than or equal to 100",
                      "type": "less_than_equal"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentialsBearer": [
              "bot:write"
            ]
          }
        ],
        "summary": "Commit Tutor Activation Endpoint",
        "tags": [
          "bot"
        ]
      }
    },
    "/api/v1/bot/tutors/{guild_id}/{tutor_discord_id}/deactivate/prepare": {
      "post": {
        "operationId": "bot_prepare_tutor_deactivation",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Guild Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "tutor_discord_id",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Tutor Discord Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionPrepareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "invalid_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Invalid authentication credentials"
                    }
                  },
                  "missing_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Not authenticated"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "code": "forbidden",
                  "detail": "Not enough permissions"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing required scope: bot:write"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "transition_conflict": {
                    "value": {
                      "code": "transition_conflict",
                      "detail": "Transition conflict"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Transition conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "transition_validation": {
                    "value": {
                      "code": "transition_validation",
                      "detail": "Transition validation failed"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Transition validation failed"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentialsBearer": [
              "bot:write"
            ]
          }
        ],
        "summary": "Prepare Tutor Deactivation Endpoint",
        "tags": [
          "bot"
        ]
      }
    },
    "/api/v1/bot/tutors/{guild_id}/{tutor_discord_id}/deactivate/{operation_id}/commit": {
      "post": {
        "operationId": "bot_commit_tutor_deactivation",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Guild Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "tutor_discord_id",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Tutor Discord Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionCommitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "invalid_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Invalid authentication credentials"
                    }
                  },
                  "missing_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Not authenticated"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "code": "forbidden",
                  "detail": "Not enough permissions"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing required scope: bot:write"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "operation_not_found": {
                    "value": {
                      "code": "operation_not_found",
                      "detail": "Operation not found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operation not found"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "operation_not_pending": {
                    "value": {
                      "code": "operation_not_pending",
                      "detail": "Operation is not in a prepared state (already committed, failed, or expired)"
                    }
                  },
                  "transition_conflict": {
                    "value": {
                      "code": "transition_conflict",
                      "detail": "Transition conflict"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operation is not in a prepared state (already committed, failed, or expired) / Transition conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "code": "validation_error",
                  "detail": "Request validation failed",
                  "errors": [
                    {
                      "loc": [
                        "query",
                        "limit"
                      ],
                      "message": "Input should be less than or equal to 100",
                      "type": "less_than_equal"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentialsBearer": [
              "bot:write"
            ]
          }
        ],
        "summary": "Commit Tutor Deactivation Endpoint",
        "tags": [
          "bot"
        ]
      }
    },
...
  },
...
}